Allow mutating candidate in `Subproblem::branch_or_evaluate` +minor
Allow candidates to mutate themselves inside
`Subproblem::branch_or_evaluate`. This is useful in case a candidate
discovers that they are a solution and wants to keep some information
about that in itself (e.g., when applying a heuristic solves the
problem, we can apply it in-place).

Also fixed a comment on `Subproblem::Score`.