Knapsack examples: both examples should use the boundary!
Fixes a bug: the knapsack-without-library example was not using
the boundary function in any way, so, rather than implementing a
branch-and-bound algorithm, it implemented a brute-force algorithm.

Moved some functions and added boundary checks
(both lazy and eager evaluation strategies) to knapsack-without-library.

Additionally, simplified the storage of the incumbent in
knapsack-without-library (we always have an incumbent).